requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
yarl==1.8.1
